+++ PayPal/src/main/java/com/braintreepayments/api/paypal/PayPalCheckoutRequest.kt
@@ -55,9 +55,12 @@ import org.json.JSONObject
  *
  * @property shouldOfferPayLater Offers PayPal Pay Later if the customer qualifies. Defaults to
  * false.
+ *
  * @property shippingCallbackUrl Server side shipping callback URL to be notified when a customer
  * updates their shipping address or options. A callback request will be sent to the merchant server
  * at this URL.
+ *
+ * @property contactInformation Contact information of the recipient for the order
  */
 @Parcelize
 class PayPalCheckoutRequest @JvmOverloads constructor(
@@ -69,6 +72,7 @@ class PayPalCheckoutRequest @JvmOverloads constructor(
     var shouldRequestBillingAgreement: Boolean = false,
     var shouldOfferPayLater: Boolean = false,
     var shippingCallbackUrl: Uri? = null,
+    var contactInformation: PayPalContactInformation? = null,
     override var localeCode: String? = null,
     override var billingAgreementDescription: String? = null,
     override var isShippingAddressRequired: Boolean = false,
@@ -134,6 +138,11 @@ class PayPalCheckoutRequest @JvmOverloads constructor(
         }
 
         userPhoneNumber?.let { parameters.put(PHONE_NUMBER_KEY, it.toJson()) }
+
+        contactInformation?.let { info ->
+            info.recipientEmail?.let { parameters.put(RECIPIENT_EMAIL_KEY, it) }
+            info.recipentPhoneNumber?.let { parameters.put(RECIPIENT_PHONE_NUMBER_KEY, it.toJson()) }
+        }
 
         if (currencyCode == null) {
             currencyCode = configuration?.payPalCurrencyIsoCode